test(backend_parity): update tests for new ops and corrected kernels
Bump head dimensions from 16 to 32 or 64 across attention, RoPE, and
paged-attention parity tests to meet CUDA kernel alignment requirements
(kernels tile on multiples of 32).

Add assert_parity_f32_relaxed and assert_parity_f32_tol helpers for
cases where backends differ in floating-point ordering (atomicAdd in
backward passes, F32→F16→FP8→F16→F32 roundtrip in FP8 KV cache quant).

Fix slot_mapping dtype in reshape_and_cache test from i64 to i32 to
match the kernel signature.

Skip wgpu backward pass tests gracefully instead of panicking; the
forward pass is tested and the backward is noted as not yet implemented.

Add fused optimizer backend parity tests covering all four optimizers
(AdamW, SGD, AdaGrad, LAMB) across CPU, CUDA, and wgpu backends.

Handle CudaClient construction failure in with_cuda helper so tests
skip cleanly on machines without a GPU rather than crashing.

+++ tests/backend_parity/kv_cache.rs
@@ -125,7 +125,7 @@ fn test_reshape_and_cache_parity() {
         &cpu_device,
     );
     // Slot mapping: tokens go into slots 0,1,4,5 (block 0 slots 0-1, block 1 slots 0-1)
-    let slot_data: Vec<i64> = vec![0, 1, 4, 5];
+    let slot_data: Vec<i32> = vec![0, 1, 4, 5];
     let slot_mapping = numr::tensor::Tensor::from_slice(&slot_data, &[num_tokens], &cpu_device);
 
     cpu_client
